pyproject.toml
[project]
name = "lacrei-saude-api"
version = "0.1.0"
description = "API RESTful para gerenciamento de consultas médicas - Lacrei Saúde"
authors = [
    {name = "Lucas Felipe dos Anjos",email = "lfanjosus@gmail.com"}
]
requires-python = ">=3.12,<4.0"
dependencies = [
    "django (>=6.0.1,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "django-filter (>=24.3,<25.0)",
    "drf-yasg (>=1.21.7,<2.0.0)",
    "djangorestframework-simplejwt (>=5.3.1,<6.0.0)",
    "django-ratelimit (>=4.1.0,<5.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
    "coverage (>=7.13.2,<8.0.0)",
    "black (>=26.1.0,<27.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "pytest-cov (>=6.0.0,<7.0.0)",
    "safety (>=3.2.0,<4.0.0)"
]

# Black configuration
[tool.black]
line-length = 127
target-version = ['py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | \.pytest_cache
  | __pycache__
  | migrations
  | staticfiles
  | media
  | htmlcov
  | dist
  | build
)/
'''

# isort configuration
[tool.isort]
profile = "black"
line_length = 127
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = [
    "*/migrations/*", 
    "*/venv/*", 
    "*/env/*", 
    "*/.pytest_cache/*",
    "*/staticfiles/*",
    "*/media/*",
    "*/htmlcov/*"
]
known_django = "django"
known_first_party = ["lacrei_saude", "authentication", "profissionais", "consultas"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "DJANGO", "FIRSTPARTY", "LOCALFOLDER"]

# Coverage configuration
[tool.coverage.run]
source = "."
branch = true
omit = [
    "*/migrations/*",
    "*/venv/*",
    "*/env/*",
    "*/test_*",
    "manage.py",
    "*/settings/*",
    "*/.*",
    "conftest.py",
    "coverage_scripts.py"
]

[tool.coverage.report]
show_missing = true
skip_covered = false
precision = 2
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "def __str__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "@abstractmethod"
]

[tool.coverage.html]
directory = "htmlcov"
title = "Lacrei Saúde API - Coverage Report"

# Pytest configuration
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "lacrei_saude.settings_test"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--tb=short", 
    "--strict-markers",
    "--reuse-db",
    "--nomigrations"
]
filterwarnings = [
    "ignore::django.utils.deprecation.RemovedInDjango50Warning",
    "ignore::DeprecationWarning"
]
testpaths = ["."]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests", 
    "unit: marks tests as unit tests",
    "models: marks tests for models",
    "views: marks tests for views",
    "serializers: marks tests for serializers",
    "validators: marks tests for validators",
    "security: marks tests for security features"
]